votes and wins functioning on the backend, with some front end

+++ controller/ControllerVote.php
@@ -3,6 +3,7 @@
 RequirePage::model('Vote');
 RequirePage::model('Writer');
 RequirePage::model('Text');
+RequirePage::controller('ControllerGame');
 
 class ControllerVote extends Controller {
 
@@ -10,32 +11,64 @@ class ControllerVote extends Controller {
 
     }
 
-    public function voteToggle($id){
-        //TODO: doublecheck if you're allowed to vote on this story
-        //STEPS: find the root parent, do a hasContributed check from root parent
-        //I can wait and do this later
-
-        //TODO: check if there's such a vote in the db
-        //STEPS: SELECT if you get an answer--unvote, else vote
+    public function voteToggle($id, $isConfirmed = false) {
         $vote = new Vote;
-
         $text_id = $id;
         $writer_id = $_SESSION['writer_id'];
-        $data = ['writer_id'=>$writer_id,'text_id'=>$text_id];
-
-        //check if theres a vote in the db for this text and writer
+        $data = ['writer_id' => $writer_id, 'text_id' => $text_id];
+    
+        // Check if this is a vote or an unvote action
         $alreadyVoted = $vote->selectCompositeId($data);
+    
+        // Get number of votes and players from db
+        $result = $vote->checkWin($text_id);
+    
+        // Adjust the vote count from the db, with the current vote
+        $voteCount = $alreadyVoted ? $result['voteCount'] - 1 : $result['voteCount'] + 1;
+        $playerCountMinusOne = $result['playerCountMinusOne'];
+    
+        // Check if the vote makes the text a winning text
+        $isWinningVote = $voteCount >= $playerCountMinusOne;
 
+        //initialize $voted to something... not true or false...
+        $voted = "pending";
+    
         // Toggle between saving or deleting the vote
-        if ($alreadyVoted) {
-            $vote->deleteVote($data);
-            $voted = false;
-        } else {
-            $vote->saveVote($data);
-            $voted = true;
+        // But first, make sure that it isn't a winning vote
+        // Or that if it is a winning vote, that it has been confirmed
+        if(!$isWinningVote || ($isWinningVote && $isConfirmed == "true")){
+            if ($alreadyVoted) {
+                $vote->deleteVote($data);
+                $voted = false;
+            } else {
+                $vote->saveVote($data);
+                $voted = true;
+            }
         }
 
+        // HERE. TRIGGER THE WIN STUFF... UPDATE THE GAME
+        if($isWinningVote && $isConfirmed == "true"){
+            $gameController = new ControllerGame;
+            $gameController->closeGame($text_id); 
+        }
+
+        //check if you are getting gameID
+        $text = new Text;
+        $gameId = $text->selectGameId($text_id); 
+    
+        // Build the response data
+        $response = [
+            'alreadyVoted' => $alreadyVoted,
+            'voted' => $voted,
+            'voteCount' => $voteCount,
+            'playerCountMinusOne' => $playerCountMinusOne,
+            'isWinningVote' => $isWinningVote,
+            'isConfirmed' => $isConfirmed,
+            'confirmationRequired' => $isWinningVote && $isConfirmed == "false",
+            'gameId' => $gameId
+        ];
+    
         // Return a JSON response
-        echo json_encode(['voted' => $voted]);
+        echo json_encode($response);
     }
 }
